% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{as.data.frame.inspectorORF_orflist}
\alias{as.data.frame.inspectorORF_orflist}
\title{Coerce the ORF object into a data.frame}
\usage{
\method{as.data.frame}{inspectorORF_orflist}(orfs)
}
\arguments{
\item{orfs}{The object created from inspectorORF::find_orfs()}
}
\value{
A data.frame consisting of the ORFs
}
\description{
Coerce the ORF object into a data.frame
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.annotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orfs <- inspectorORF::find_orfs(test_tracks, "ENST00000643391.1")
df <- test_orfs |> as.data.frame()
}
