% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{codon_query}
\alias{codon_query}
\title{Used to annotate codons on the ORF plots.}
\usage{
codon_query(
  annotation_codons = NULL,
  annotate_start = F,
  in_frame = F,
  annotate_stop = F,
  colour = NULL
)
}
\arguments{
\item{annotation_codons}{a vector of strings consisting of which codons to search for within the sequence being plot}

\item{annotate_start}{a logical indicating if the start codon of the plotted ORF should be anotated}

\item{in_frame}{a logical indicating if the annotation_codons should be in frame to the ORF or not}

\item{annotate_stop}{if true and one of annotation_codons or annotate_start is included along with in_frame = TRUE, stop codons will also be annotated. Can be logical value to plot all three stop codons, or vector of strings to specify codons.}

\item{colour}{a string consisting of the colour to plot the annotated codons in}
}
\value{
a named list of set codon queries for use in the codon_queries option of the orf_plot function
}
\description{
A simple wrapper function for a named list to specify codon types you wish to annotate on the ORF/transcript plot
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
test_orf_plot <- inspectorORF::orf_plot(
  tracks,
  transcript_filter = "ENST00000343702.9",
  start_position = 456,
  codon_queries = list(codon_query(annotate_start = TRUE))
)
test_orf_plot
}
