% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{find_orfs}
\alias{find_orfs}
\title{Search for ORFs found within the tracks of a transcript}
\usage{
find_orfs(
  transcript_tracks,
  transcript_filter,
  frame_filter = c(0, 1, 2),
  start_threshold_filter = 13,
  start_codon_filter = c("AUG", "CUG", "GUG", "UUG", "AAG", "ACG"),
  stop_codon_filter = c("UAG", "UGA", "UAA"),
  orf_length_filter = 3
)
}
\arguments{
\item{transcript_tracks}{The object created from one of the import functions}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{frame_filter}{The frame(s) to search in (optional) (defailt: 0, 1 and 2)}

\item{start_threshold_filter}{The minimum number of nucleotides the start site is to be looked at (optional) (default: 17)}

\item{start_codon_filter}{The start codons to be used for ORF searching (optional) (default: "AUG", "CUG", "GUG", "UUG", "AAG", and "ACG")}

\item{stop_codon_filter}{The stop codons to be used to determine the end of the ORF (optional) (default: "UAG", "UGA", and "UAA")}

\item{orf_length_filter}{The minimum length to be considered an ORF (optional) (default: 3)}
}
\value{
A tibble of ORFs consisting of frame, start codon, stop codon, start position, stop position, length in nucleotides and length in amino acids
}
\description{
Creates a table of ORFs found within a transcript, where ORF is defined as
a start (default "AUG", "CUG", "GUG", "UUG", "AAG", and "ACG") and stop ("UAG", "UGA", or "UAA").
This does not perform any checks to see if the ORF is translated, it is simply an in-frame start and stop search.
}
\examples{
tracks <- inspectorORF::import_orf_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orfs <- inspectorORF::find_orfs(test_tracks, "ENST00000643391.1")
test_orfs
}
