% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{gene_to_transcript_tracks}
\alias{gene_to_transcript_tracks}
\title{Extracts reads from gene level data relevant to a transcript of interest}
\usage{
gene_to_transcript_tracks(gene_tracks, transcript_id_filter)
}
\arguments{
\item{gene_tracks}{The gene tracks object obtained from inspectorORF::import_gene_tracks}

\item{transcript_id_filter}{The transcript ids of interest}
}
\value{
A transcript tracks object
}
\description{
Extracts reads from gene level data relevant to a transcript of interest
}
\examples{
gene_tracks <- inspectorORF::import_gene_tracks("test.bed", "gencode.v44.annotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
tx_tracks <- inspectorORF::gene_to_transcript_tracks(gene_tracks, transcript_id_filter = c("ENST00000361764.9", "ENST00000643391.1"))
}
