% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{get_gene_tracks}
\alias{get_gene_tracks}
\title{Obtain tracks for a specific gene or genes from RiboTaper or ORFQuant results}
\usage{
get_gene_tracks(
  tracks,
  gtf_file,
  genome_file,
  gene_ids,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{tracks}{the tracks object consisting of}

\item{gtf_file}{Path to the gtf annotation}

\item{gene_ids}{}
}
\value{
A GRangesList of all the genes requested
}
\description{
Obtain tracks for a specific gene or genes from RiboTaper or ORFQuant results
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant")

tx_tracks <- get_gene_tracks("annotation.gtf", gene_ids = c("ENSG000005234.1"))
}
