% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{get_id_tracks}
\alias{get_id_tracks}
\title{Obtain tracks for a set of genes or transcript tracks by identifying the type of ids specified automatically}
\usage{
get_id_tracks(
  tracks,
  gtf_file,
  genome_file,
  ids,
  keep_introns = FALSE,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{tracks}{the tracks object consisting of}

\item{gtf_file}{Path to the gtf annotation}

\item{genome_file}{Path to the genome annotation 2bit file}

\item{ids}{the gene ids of interest. If empty, will default to returning transcript tracks}

\item{keep_introns}{should introns be retained during generation of tracks}

\item{framed_tracks}{the name of rows which contain p-site data. Defaults to p_sites.}
}
\value{
An inspectorORF gene or transcript tracks object of all the ids requested
}
\description{
Obtain tracks for a set of genes or transcript tracks by identifying the type of ids specified automatically
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = system.file(
    "example_data", "control_rna_tracks.bed.gz", package = "inspectorORF"
  ),
  orfquant_psites = system.file(
    "example_data", "control_psites_for_ORFquant", package = "inspectorORF"
  )
)

gene_tracks <- inspectorORF::get_id_tracks(
  tracks,
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
  ids = c("ENSG00000074527.13")
)

}
