% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf_aa_seq}
\alias{get_orf_aa_seq}
\title{Extract the amino acid seqeunce of an ORF within a transcript}
\usage{
get_orf_aa_seq(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter,
  start_position = NULL,
  stop_position = NULL
)
}
\arguments{
\item{transcript_tracks}{The object created from inspectorORF::create_tracks()}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}
}
\value{
An AAStringSet consisting of the ORF amino acid residues
}
\description{
Extracts the ORF sequence from a transcript track,
if the stop position is omitted, the nearest downstream, in-frame stop codon is used from the supplied start codon
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
orf_aa_sequence <- inspectorORF::get_orf_aa_seq(tracks, transcript_filter = "ENST00000486256.5", start_position = 1702)
orf_aa_sequence
}
