% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf_framing}
\alias{get_orf_framing}
\title{Obtain reads for all three frames of an ORF}
\usage{
get_orf_framing(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL
)
}
\arguments{
\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}

\item{inspectorORF_object}{The object created from inspectorORF::create_tracks()}
}
\value{
a table consisting of summarised reads for each frame
}
\description{
Obtain the RNA-Seq and Ribo-Seq reads for all three frames from the specified ORF
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
orf_aa_sequence <- inspectorORF::get_orf_aa_seq(test_tracks, transcript_filter = "ENST00000486256.5", start_position = 1702)
orf_aa_sequence
}
