% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf_nt_seq}
\alias{get_orf_nt_seq}
\title{Extract the nucleotide seqeunce of an ORF within a transcript}
\usage{
get_orf_nt_seq(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL,
  stop_codons = c("UAG", "UAA", "UGA")
)
}
\arguments{
\item{transcript_tracks}{The object created from inspectorORF::create_tracks()}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}
}
\value{
An RNAStringSet consisting of the ORF nucleotides
}
\description{
Extracts the ORF sequence from a transcript track,
if the stop position is omitted, the nearest downstream, in-frame stop codon is used from the supplied start codon
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orfs <- inspectorORF::find_orfs(tracks, "ENST00000643391.1")
inspectorORF::get_orf_nt_seq(tracks, test_orfs[[1]])
}
