% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{get_transcript_tracks}
\alias{get_transcript_tracks}
\title{Obtain track information for one or more transcripts from RiboTaper/ORFquant results}
\usage{
get_transcript_tracks(
  tracks,
  gtf_file,
  genome_file,
  transcript_ids = NULL,
  additional_info = NULL,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{tracks}{Track information in GRanges created by import_tracks_from_RiboTaper or merge_RNA_tracks_with_ORFquant}

\item{gtf_file}{Path to a GTF annotation file}

\item{genome_file}{path to a Genome 2bit file}

\item{transcript_ids}{vector of transcript_id(s) to obtain. If NULL, all transcript_ids within the annotation will be used.}

\item{additional_info}{additional information in GRanges format with transcript level coordinates such as Kozak scores, Mass-spec data etc}

\item{framed_tracks}{names consisting of p_sites. e.g, if there are two conditions, both with ORFquant P-sites read in, this would}
}
\value{
An inspectorORF transcript tracks object
}
\description{
Obtain track information for one or more transcripts from RiboTaper/ORFquant results
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant")

tx_tracks <- get_transcript_tracks("annotation.gtf")
}
