% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{get_transcripts_for_gene}
\alias{get_transcripts_for_gene}
\title{Get a vector of all transcript ids found within a particular gene_id}
\usage{
get_transcripts_for_gene(gene_tracks, gene_filter)
}
\arguments{
\item{gene_tracks}{The gene tracks object, obtained from import_gene_tracks}

\item{gene_filter}{The gene_id of interest}
}
\value{
A vector of transcript_ids associated with the gene_id
}
\description{
Get a vector of all transcript ids found within a particular gene_id
}
\examples{
gene_tracks <- inspectorORF::import_gene_bed(
  system.file("example_data", "example_gene.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF")
)
print(inspectorORF::get_transcripts_for_gene(
  gene_tracks,
  gene_filter = "ENSG00000074527.1"
))
}
