% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_gene.R
\name{import_gene_bed}
\alias{import_gene_bed}
\title{Import reads from a bed file where the tracks are gene level information}
\usage{
import_gene_bed(bed_file, gtf_file, genome_file, framed_tracks = c("p_sites"))
}
\arguments{
\item{bed_file}{The bed file containing the gene tracks, each track must contain the trackline of gene_id}

\item{gtf_file}{The path to the gtf file}

\item{genome_file}{The path to the genome fasta or 2bit file}

\item{framed_tracks}{The track line consisting of the P-site reads (defaults to p_sites)}
}
\value{
An inspectorORF_gtracks object consisting of the relevant reads
}
\description{
Import reads from a bed file where the tracks are gene level information
}
\examples{
gene_tracks <- inspectorORF::import_gene_bed(
  system.file("example_data", "example_gene.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF")
)
}
