% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{import_orf_tracks}
\alias{import_orf_tracks}
\title{Import reads from a bed file where the tracks are ORF level information}
\usage{
import_orf_tracks(
  bed_file,
  gtf_file,
  genome_file,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{bed_file}{The bed file containing the ORF tracks, each track must contain the trackline of transcript_id_start_stop, e.g ENST00000361764.9_236_400}

\item{gtf_file}{The path to the gtf file}

\item{genome_file}{The path to the genome fasta or 2bit file}

\item{framed_tracks}{The track line consisting of the P-site reads (defaults to p_sites)}
}
\value{
An inspectorORF_txtracks object consisting of the relevant reads
}
\description{
Import reads from a bed file where the tracks are ORF level information
}
\examples{
tracks <- inspectorORF::import_orf_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
}
