% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{import_tracks_from_RiboTaper}
\alias{import_tracks_from_RiboTaper}
\title{Imports RNA-Seq and P-Site reads from RiboTaper results}
\usage{
import_tracks_from_RiboTaper(...)
}
\arguments{
\item{...}{Read from one or more RiboTaper result files ('Psit_Ribo_Rna_Cent_tracks_ccds' and 'Psit_Ribo_Rna_Cent_tracks_nonccds'), currently does not support multiple conditions or replicates}
}
\value{
returns a GRanges object
}
\description{
Imports RNA-Seq and P-Site reads from RiboTaper results
}
\examples{
tracks <- inspectorORF::import_tracks_from_RiboTaper(
  c('Psit_Ribo_Rna_Cent_tracks_ccds', 'Psit_Ribo_Rna_Cent_tracks_nonccds')
)
}
