% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{import_transcript_bed}
\alias{import_transcript_bed}
\title{Import reads from a bed file where the tracks are transcript level information}
\usage{
import_transcript_bed(
  bed_file,
  gtf_file,
  genome_file,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{bed_file}{The bed file containing the ORF tracks, each track must contain the trackline of a transcript_id}

\item{gtf_file}{The path to the gtf file}

\item{genome_file}{The path to the genome fasta or 2bit file}

\item{framed_tracks}{The track line consisting of the P-site reads (defaults to p_sites)}
}
\value{
An inspectorORF_txtracks object consisting of the relevant reads
}
\description{
Import reads from a bed file where the tracks are transcript level information
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.annotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
}
