% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{merge_RNA_tracks_with_ORFquant}
\alias{merge_RNA_tracks_with_ORFquant}
\title{Create a tracks object from quantified RNA reads and P-site information from ORFquant}
\usage{
merge_RNA_tracks_with_ORFquant(
  rna_reads,
  rna_abundance = NULL,
  orfquant_psites,
  extra = NULL,
  extra_orfquant_psites = NULL,
  orfquant_results_type = "P_sites_all"
)
}
\arguments{
\item{rna_reads}{The quantified RNA-Seq reads in bed format, typically generated by bedtools}

\item{rna_abundance}{An optional quantified abundance RNA sites in bed format, typically generated by bedtools (see inspectorORF vignette)}

\item{orfquant_psites}{The _for_ORFquant file generated by ORFquant during P-site calculation}

\item{extra}{Additional optional reads in bed format such as a second condition, additional replicates, or additional data such as long-read sequencing (as a named list)}

\item{extra_orfquant_psites}{Additional optional ORFquant-calculated P-sites for additional conditions or replicates (as a named list)}

\item{orfquant_results_type}{Which set of P-site information from the ORFquant data such as P_sites_all or P_sites_uniq. Defaults to P_sites_all}
}
\value{
A GRanges object
}
\description{
Create a tracks object from quantified RNA reads and P-site information from ORFquant
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant")

# Creating a tracks object with multiple conditions
# with the control under the regular rna_reads/orfquant_psites and an additional treated label
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant",
  extra = list(treated_rna_reads = "treated_rna_reads.bed"),
  extra_orfquant_psites = list(treated_psites = "treated_reads_for_ORFquant"))

# Creating a tracks object with multiple conditions and additional read data for control
# with the control under the regular rna_reads/orfquant_psites and an additional treated label
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant",
  extra = list(treated_rna_reads = "treated_rna_reads.bed",
               dRNA_reads = "dRNA_reads.bed"),
  extra_orfquant_psites = list(treated_psites = "treated_reads_for_ORFquant"))
}
