% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{orf_plot}
\alias{orf_plot}
\title{Plot the reads within an ORF of a transcript}
\usage{
orf_plot(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL,
  plot_region = c(start_position, stop_position),
  plot_colours = c(rna_reads = "grey60", `0` = "#440854", `1` = "#23A884", `2` =
    "#FEE725"),
  scale_to_psites = F,
  summarise_outside_orf = F,
  codon_queries = NULL,
  condition_names = c(rna_reads = ""),
  plot_read_pairs = c(p_sites = "rna_reads"),
  dataset_names = c(rna_reads = "RNA-Seq Reads", p_sites = "P-Sites"),
  interactive = F,
  one_plot = T,
  legend_position = "bottom",
  text_size = 12
)
}
\arguments{
\item{transcript_tracks}{The object created from importing the tracks}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}

\item{plot_region}{an optional tuple consisting of a region of 5'utr and 3'utr to plot}

\item{plot_colours}{The custom colour scheme to use for frame 0, 1 and 2 (optional)}

\item{scale_to_psites}{Should the plot be scaled to the highest P-site peak, therefore cutting off any RNA-Seq reads above this}

\item{summarise_outside_orf}{Plots a summary of P-site reads for region outside the ORF}

\item{codon_queries}{a list consisting of one or more inspectorORF::codon_queries() calls, indiciating any codons to be annotated within the plot}

\item{condition_names}{Names of any datasets to plot, useful when plotting bed file which consists of reads from multiple datasets}

\item{plot_read_pairs}{Which RNA-Seq reads are associated with which P-Site reads. See example for further info}

\item{dataset_names}{Custom naming to display on the plot for any read type}

\item{interactive}{Enable to return an interactive plotly figure}

\item{one_plot}{Should a combined figure be returned or individual figures for main plot and any triplet periodicity plots}

\item{legend_position}{Location of the legend within the figure}

\item{text_size}{Text size within the figure}
}
\value{
A ggplot of the ORF
}
\description{
Plot the reads within an ORF of a transcript
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
test_orf_plot <- inspectorORF::orf_plot(
  tracks,
  transcript_filter = "ENST00000343702.9",
  start_position = 456
)
test_orf_plot
}
