% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_shiny.R
\name{run_inspectorORF}
\alias{run_inspectorORF}
\title{Launch an interactive application of inspectorORF}
\usage{
run_inspectorORF(
  launch = TRUE,
  host = "127.0.0.1",
  port = 4739,
  launch_browser = TRUE
)
}
\arguments{
\item{launch}{Run immediately (default) or just return the shiny app object}

\item{host}{Host to start the shiny app on (defaults to 127.0.0.1)}

\item{port}{Port the shiny app will listen on (defaults to 4739)}

\item{launch_browser}{launch the browser automatically if TRUE (default)}
}
\value{
a shiny object if launch = FALSE, otherwise, runs the app
}
\description{
Build and optionally launch the main application
}
