% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{save_tracks}
\alias{save_tracks}
\title{Saves the merged RNA and P-site read information into a bed file}
\usage{
save_tracks(tracks, filename)
}
\arguments{
\item{tracks}{the tracks object created by import_tracks_from_RiboTaper or merge_RNA_tracks_with_ORFquant}

\item{filename}{file path to save the tracks bed file to}
}
\description{
Saves the merged RNA and P-site read information into a bed file
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = system.file("example_data", "control_rna_tracks.bed.gz", package = "inspectorORF"),
  orfquant_psites = system.file(
    "example_data", "control_psites_for_ORFquant", package = "inspectorORF"
  )
)

tx_tracks <- inspectorORF::get_transcript_tracks(
  tracks,
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
  transcript_ids = c("ENST00000343702.9")
)
save_tracks(tx_tracks, "output_tracks.bed")
}
