% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{as.data.frame.inspectorORF_orflist}
\alias{as.data.frame.inspectorORF_orflist}
\title{Coerce the ORF object into a data.frame}
\usage{
\method{as.data.frame}{inspectorORF_orflist}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{The object created from inspectorORF::find_orfs()}

\item{row.names}{unused}

\item{optional}{unused}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A data.frame consisting of the ORFs
}
\description{
Coerce the ORF object into a data.frame
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
test_orfs <- inspectorORF::find_orfs(
  tracks,
  "ENST00000343702.9"
)
df <- test_orfs |> as.data.frame()
}
