% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{codon_query}
\alias{codon_query}
\title{Used to annotate codons on the ORF plots.}
\usage{
codon_query(
  annotation_codons = NULL,
  annotate_start = F,
  in_frame = F,
  colour = NULL
)
}
\arguments{
\item{annotation_codons}{a vector of strings consisting of which codons to search for within the sequence being plot}

\item{annotate_start}{a logical indicating if the start codon should be anotated}

\item{in_frame}{a logical indicating if the annotation_codons should be in frame to the ORF or not}

\item{colour}{a string consisting of the colour to plot the annotated codons in}
}
\value{
a named list
}
\description{
A simple wrapper function for a named list to specify codon types you wish to annotate on the ORF/transcript plot
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.annotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orf_plot <- inspectorORF::orf_plot(tracks, transcript_filter = "ENST00000486256.5", start_position = 1702, codon_queries = list(codon_query(annotate_start = T)))
test_orf_plot
}
