% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{gene_to_transcript_tracks}
\alias{gene_to_transcript_tracks}
\title{Extracts reads from gene level data relevant to a transcript of interest}
\usage{
gene_to_transcript_tracks(gene_tracks, transcript_filter)
}
\arguments{
\item{gene_tracks}{The gene tracks object obtained from inspectorORF::import_gene_tracks}

\item{transcript_filter}{The transcript ids of interest}
}
\value{
A transcript tracks object
}
\description{
Extracts reads from gene level data relevant to a transcript of interest
}
\examples{
gene_tracks <- inspectorORF::import_gene_bed(
  system.file("example_data", "example_gene.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF")
)
tx_tracks <- inspectorORF::gene_to_transcript_tracks(
  gene_tracks,
  transcript_filter = c("ENST00000343702.9", "ENST00000344911.8")
)
}
