% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{get_gene_tracks}
\alias{get_gene_tracks}
\title{Obtain tracks for a specific gene or genes from RiboTaper or ORFQuant results}
\usage{
get_gene_tracks(
  tracks,
  gtf_file,
  genome_file,
  gene_ids,
  framed_tracks = c("p_sites")
)
}
\arguments{
\item{tracks}{the tracks object consisting of}

\item{gtf_file}{Path to the gtf annotation}

\item{genome_file}{Path to the genome annotation 2bit file}

\item{gene_ids}{the gene ids of interest}

\item{framed_tracks}{the name of rows which contain p-site data. Defaults to p_sites.}
}
\value{
An inspectorORF gene tracks object of all the genes requested
}
\description{
Obtain tracks for a specific gene or genes from RiboTaper or ORFQuant results
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = system.file(
    "example_data", "control_rna_tracks.bed.gz", package = "inspectorORF"
  ),
  orfquant_psites = system.file(
    "example_data", "control_psites_for_ORFquant", package = "inspectorORF"
  )
)

gene_tracks <- inspectorORF::get_gene_tracks(
  tracks,
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
  gene_ids = c("ENSG00000074527.13")
)

}
