% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf}
\alias{get_orf}
\title{Get a specific ORF}
\usage{
get_orf(orf_object, start_position)
}
\arguments{
\item{orf_object}{The object created from inspectorORF::find_orfs()}

\item{start_position}{The position of the start site of interest}
}
\value{
A list containing information about the orfs frame, start and stop codon, and length, or NULL if no ORFs were found
}
\description{
Gets details of an ORF found within the ORF list based on the start site.
This does not perform any checks to see if the ORF is translated, it is simply an in-frame start and stop search.
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orfs <- inspectorORF::find_orfs(test_tracks, "ENST00000643391.1")
orf <- inspectorORF::get_orf(test_orfs, start_position = 49)
orf
}
