% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf_aa_seq}
\alias{get_orf_aa_seq}
\title{Extract the amino acid seqeunce of an ORF within a transcript}
\usage{
get_orf_aa_seq(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter,
  start_position = NULL,
  stop_position = NULL,
  stop_codons = c("UAG", "UAA", "UGA")
)
}
\arguments{
\item{transcript_tracks}{The object created from inspectorORF::create_tracks()}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}

\item{stop_codons}{Which stop codons to consider end of the ORF sequence, defaults to UAG, UAA and UGA}
}
\value{
An AAStringSet consisting of the ORF amino acid residues
}
\description{
Extracts the ORF sequence from a transcript track,
if the stop position is omitted, the nearest downstream, in-frame stop codon is used from the supplied start codon
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
orf_aa_sequence <- inspectorORF::get_orf_aa_seq(
  tracks,
  transcript_filter = "ENST00000343702.9",
  start_position = 456
)
orf_aa_sequence
}
