% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{get_orf_framing}
\alias{get_orf_framing}
\title{Obtain reads for all three frames of an ORF}
\usage{
get_orf_framing(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL
)
}
\arguments{
\item{transcript_tracks}{the inspectorORF tracks object generated by import_transcript_bed or get_transcript_tracks}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}
}
\value{
a table consisting of summarised reads for each frame
}
\description{
Obtain the RNA-Seq and Ribo-Seq reads for all three frames from the specified ORF
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
orf_aa_sequence <- inspectorORF::get_orf_framing(
  tracks,
  transcript_filter = "ENST00000343702.9",
  start_position = 456
)
orf_aa_sequence
}
