% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{get_transcripts_for_gene}
\alias{get_transcripts_for_gene}
\title{Get a vector of all transcript ids found within a particular gene_id}
\usage{
get_transcripts_for_gene(gene_tracks, gene_id_filter)
}
\arguments{
\item{gene_tracks}{The gene tracks object, obtained from import_gene_tracks}

\item{gene_id_filter}{The gene_id of interest}
}
\value{
A vector of transcript_ids associated with the gene_id
}
\description{
Get a vector of all transcript ids found within a particular gene_id
}
\examples{
gene_tracks <- inspectorORF::import_gene_tracks("test.bed", "gencode.v44.annotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
print(inspectorORF::get_transcripts_for_gene(gene_tracks, gene_id = "ENSG00000092096.1"))
}
