% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{orf_plot}
\alias{orf_plot}
\title{Plot the reads within an ORF of a transcript}
\usage{
orf_plot(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL,
  plot_region = c(start_position, stop_position),
  plot_colours = c(rna_reads = "grey60", `0` = "#440854", `1` = "#23A884", `2` =
    "#FEE725"),
  scale_to_psites = F,
  plot_transcript_summary = F,
  codon_queries = NULL,
  condition_names = c(rna_reads = ""),
  plot_read_pairs = c(p_sites = "rna_reads"),
  dataset_names = c(rna_reads = "RNA-Seq Reads", p_sites = "P-Sites"),
  one_plot = T,
  legend_position = "bottom",
  text_size = 12
)
}
\arguments{
\item{transcript_tracks}{The object created from importing the tracks}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start position to use for the ORF}

\item{stop_position}{The stop position to use for the ORF (optional)}

\item{plot_region}{an optional tuple consisting of a region of 5'utr and 3'utr to plot}

\item{plot_colours}{The custom colour scheme to use for frame 0, 1 and 2 (optional)}

\item{scale_to_psites}{Should the plot be scaled to the highest P-site peak, therefore cutting off any RNA-Seq reads above this}

\item{plot_transcript_summary}{Should the remaining P-sites for the full transcript (excluding those in the ORF) be plot}

\item{codon_queries}{a list consisting of one or more inspectorORF::codon_queries() calls, indiciating any codons to be annotated within the plot}

\item{condition_names}{Names of any datasets to plot, useful when plotting bed file which consists of reads from multiple datasets}

\item{plot_read_pairs}{Which RNA-Seq reads are associated with which P-Site reads. See example for further info}

\item{dataset_names}{Custom naming to display on the plot for any read type}
}
\value{
A ggplot of the ORF
}
\description{
Plot the reads within an ORF of a transcript
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orf_plot <- inspectorORF::orf_plot(tracks, transcript_filter = "ENST00000486256.5", start_position = 1702)
test_orf_plot
}
