% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_orf.R
\name{print.inspectorORF_orflist}
\alias{print.inspectorORF_orflist}
\title{Print obtained ORFs}
\usage{
\method{print}{inspectorORF_orflist}(orf, ..., quote = FALSE)
}
\arguments{
\item{orf}{The object created from inspectorORF::find_orfs()}
}
\description{
Prints out the obtained ORFs as a data.frame
}
\examples{
tracks <- inspectorORF::import_transcript_tracks("test.bed", "gencode.v44.nnotation.gtf", "GRCh38.p14.primary_assembly.genome.2bit")
test_orfs <- inspectorORF::find_orfs(test_tracks, "ENST00000643391.1")
print(test_orfs)
}
