% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_processing.R
\name{save_tracks}
\alias{save_tracks}
\title{Saves the merged RNA and P-site read information into a bed file}
\usage{
save_tracks(tracks, filename)
}
\arguments{
\item{tracks}{the tracks object created by import_tracks_from_RiboTaper or merge_RNA_tracks_with_ORFquant}

\item{filename}{file path to save the tracks bed file to}
}
\description{
Saves the merged RNA and P-site read information into a bed file
}
\examples{
tracks <- inspectorORF::merge_RNA_tracks_with_ORFquant(
  rna_reads = "rna_reads.bed",
  orfquant_psites = "reads_for_ORFquant")

tx_tracks <- get_transcript_tracks("annotation.gtf")
save_tracks(tx_tracks, "output_tracks.bed")
}
