% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectorORF_transcript.R
\name{transcript_plot}
\alias{transcript_plot}
\title{Plot the reads for a full transcript}
\usage{
transcript_plot(
  transcript_tracks,
  orf_object = NULL,
  transcript_filter = NULL,
  start_position = NULL,
  stop_position = NULL,
  plot_colours = c(rna_reads = "grey60", `0` = "#440854", `1` = "#23A884", `2` =
    "#FEE725"),
  scale_to_psites = F,
  split_exons = F,
  plot_transcript_summary = F,
  codon_queries = NULL,
  condition_names = c(rna_reads = ""),
  plot_read_pairs = c(p_sites = "rna_reads"),
  dataset_names = c(rna_reads = "RNA-Seq Reads", p_sites = "P-Sites"),
  one_plot = T,
  interactive = F,
  legend_position = "bottom",
  text_size = 12
)
}
\arguments{
\item{transcript_tracks}{The object created from inspectorORF::create_tracks()}

\item{orf_object}{an ORF object returned by inspectorORF::get_orf() - Optional, used instead of transcript_filter, start_position and stop_position}

\item{transcript_filter}{A character vector consisting of your transcript ID of interest}

\item{start_position}{The start location of the ORF to highlight (optional)}

\item{stop_position}{The stop location of the ORF to highlight, if omitted, the nearest downstream, in-frame stop codon from the supplied start position will be used.}

\item{plot_colours}{The colour scheme for frame 0, 1 and 2 (optional).}

\item{scale_to_psites}{Should the plot be scaled to the highest P-site peak, therefore cutting off any RNA-Seq reads above this}

\item{split_exons}{separate regions by exon if TRUE, include plotting of intron regions if set to "with_introns" assuming keep_introns was set to TRUE when generating tracks. - not compatible with adding ORF regions}

\item{plot_transcript_summary}{Should the remaining P-sites for the full transcript (excluding those in the ORF) be plot}

\item{codon_queries}{an optional list consisting of one or more inspectorORF::codon_queries() calls, indiciating any codons to be annotated within the plot}

\item{condition_names}{Names of any datasets to plot, useful when plotting bed file which consists of reads from multiple datasets}

\item{plot_read_pairs}{Which RNA-Seq reads are associated with which P-Site reads. See example for further info}

\item{dataset_names}{Custom naming to display on the plot for any read type}

\item{one_plot}{Should a combined figure be returned or individual figures for main plot and any triplet periodicity plots}

\item{interactive}{Enable to return an interactive plotly figure}

\item{legend_position}{Location of the legend within the figure}

\item{text_size}{Text size within the figure}
}
\value{
a ggplot object.
}
\description{
Plots the Ribo-Seq, RNA-Seq, and Mass-Spec data across a full transcript.
Optionally, with an ORF of interest outlined.
}
\examples{
tracks <- inspectorORF::import_transcript_bed(
  system.file("example_data", "example_tracks.bed", package = "inspectorORF"),
  gtf_file = system.file("example_data", "annotation_subset.gtf", package = "inspectorORF"),
  genome_file = system.file("example_data", "chr12.2bit", package = "inspectorORF"),
)
tx_plot <- inspectorORF::transcript_plot(
  tracks,
  transcript_filter = "ENST00000343702.9"
)

# Picking out an ORF of interest within the transcript plot
tx_plot <- inspectorORF::transcript_plot(
  tracks,
  transcript_filter = "ENST00000343702.9",
  start_position = 10,
  stop_position = 30
)
}
